Doc

Documentation

data Vendor

Vendor of a docset

Constructors

DevDocs
Docset provided by devdocs.io

Hoogle
Docset provided by an archive, from which a Hoogle database can be generated

newtype Collection

To which language or library does this entry belong

Constructors

Collection

Fields

getCollection :: String

newtype Version

Version of the collection

For DevDocs, this is the version of the language or the library, e.g. the 3.6.4 as in Python 3.6.4, or the 1.13.0 as in NumPy 1.13.0.

For Hoogle, this is the NAME parameter in the doc-browser --install-hoogle command

Constructors

Version

Fields

getVersion :: String

combineCollectionVersion :: Collection -> Version -> String

breakCollectionVersion :: String -> (Collection, Version)