Utils

Documentation

(|>) :: t1 -> (t1 -> t2) -> t2 infixl 0

download :: String -> IO ByteString

downloadJSON :: FromJSON a => String -> IO a

downloadFile :: String -> Path a File -> IO ()

report :: [String] -> IO ()

updateTMVar :: TMVar a -> a -> STM ()

unpackXzInto :: Path a File -> Path a Dir -> IO ()

unpackTgzInto :: ByteString -> Path Abs Dir -> IO ()

systemUnpackTgzInto :: Path Abs File -> Path Abs Dir -> IO ()

fireAndForget :: IO a -> IO ()

localTime :: IO String

uppercaseFirst :: String -> String

lowercaseFirst :: String -> String

paragraph :: [String] -> String

paragraphs :: [[String]] -> [String]

joinDir :: Path a Dir -> [Path Rel Dir] -> Path a Dir

hasExtension :: MonadThrow m => Path a File -> String -> m Bool

newtype ConfigRoot

Constructors

ConfigRoot

Fields

getConfigRoot :: Path Abs Dir

newtype CacheRoot

Constructors

CacheRoot

Fields

getCacheRoot :: Path Abs Dir

tryRemoveFile :: Path a File -> IO ()

tryRemoveDir :: Path a Dir -> IO ()

withTempFilePath :: String -> (Path Abs File -> IO a) -> IO a

extractAp :: (a -> b -> IO c) -> IO a -> b -> IO c

reportExceptions :: SomeException -> IO ()

mapLeft :: (a -> c) -> Either a b -> Either c b

The mapLeft function takes a function and applies it to an Either value iff the value takes the form Left _.

Using Data.Bifunctor:

mapLeft = first

Using Control.Arrow:

mapLeft = (left)

Using Control.Lens:

mapLeft = over _Left

>>> mapLeft (*2) (Left 4)
Left 8

>>> mapLeft (*2) (Right "hello")
Right "hello"

withLock :: Path Abs Dir -> IO a -> IO a

(<//>) :: FilePath -> FilePath -> FilePath

(<..>) :: FilePath -> String -> FilePath