Doc

Documentation

data Vendor

Vendor of a docset

Constructors

DevDocs
Docset provided by devdocs.io

Dash
Docset provided by kapeli.com/dash

Hoogle
Docset provided by an archive, from which a Hoogle database can be generated

data Collection

Collection of the entry

For DevDocs and Dash, this is the language or library that this entry belongs to.

For Hoogle, this is the COLLECTION parameter in the doc-browser --install-hoogle command.

getCollection :: Collection -> String

parseCollection :: MonadThrow m => String -> m Collection

collection :: QuasiQuoter

newtype Version

Version of the collection

For DevDocs, this is the version of the language or the library, e.g. the 3.6.4 as in Python 3.6.4, or the 1.13.0 as in NumPy 1.13.0.

For Dash, part of this is the version of the language or the library.

For Hoogle, this is the version of the package as specified in their cabal file.

Constructors

Version

Fields

getVersion :: String

combineCollectionVersion :: Collection -> Version -> String

breakCollectionVersion :: MonadThrow m => Path a Dir -> m (Collection, Version)

data InvalidCollection

hasCVSep :: [Char] -> Bool